refactor onMethodCall into multiple classes
