+++ packages/rook_sdk_health_connect/android/src/main/kotlin/com/rookmotion/rook_sdk_health_connect/Constants.kt
@@ -1,3 +1,7 @@
 package com.rookmotion.rook_sdk_health_connect
 
+import io.flutter.plugin.common.MethodChannel
+
 const val DEFAULT_INT = 0
+
+typealias MethodResult = MethodChannel.Result
