+++ packages/rook_sdk_health_connect/android/src/main/kotlin/com/rookmotion/rook_sdk_health_connect/extension/MethodResultCommonExtensions.kt
@@ -1,0 +1,13 @@
+package com.rookmotion.rook_sdk_health_connect.extension
+
+import io.flutter.plugin.common.MethodChannel
+
+fun MethodChannel.Result.intSuccess(int: Int) {
+    success(int)
+}
+
+fun MethodChannel.Result.throwable(throwable: Throwable) {
+    error(GENERIC_ERROR_CODE, throwable.getErrorMessage(), null)
+}
+
+private const val GENERIC_ERROR_CODE = "-1"
